Enable smart weight loading for all task
