+++ src/otx/core/model/entity/base.py
@@ -5,8 +5,9 @@
 
 from __future__ import annotations
 
+import logging
 from abc import abstractmethod
-from typing import Any, Generic
+from typing import TYPE_CHECKING, Any, Generic
 
 from torch import nn
 
@@ -15,6 +16,9 @@ from otx.core.data.entity.base import (
     T_OTXBatchDataEntity,
     T_OTXBatchPredEntity,
 )
+
+if TYPE_CHECKING:
+    import torch
 
 
 class OTXModel(nn.Module, Generic[T_OTXBatchDataEntity, T_OTXBatchPredEntity]):
@@ -57,3 +61,54 @@ class OTXModel(nn.Module, Generic[T_OTXBatchDataEntity, T_OTXBatchPredEntity]):
             if self._customize_outputs != OTXModel._customize_outputs
             else outputs
         )
+
+    def register_load_state_dict_pre_hook(self, model_classes: list[str], ckpt_classes: list[str]) -> None:
+        """Register load_state_dict_pre_hook.
+
+        Args:
+            model_classes (list[str]): Class names from training data.
+            ckpt_classes (list[str]): Class names from checkpoint state dictionary.
+        """
+        self.model_classes = model_classes
+        self.ckpt_classes = ckpt_classes
+        self._register_load_state_dict_pre_hook(self.load_state_dict_pre_hook)
+
+    def load_state_dict_pre_hook(self, state_dict: dict[str, torch.Tensor], prefix: str, *args, **kwargs) -> None:
+        """Modify input state_dict according to class name matching before weight loading."""
+        model2ckpt = self.map_class_names(self.model_classes, self.ckpt_classes)
+        logger = logging.getLogger()
+        logger.info(
+            f"Data classes from checkpoint: {self.ckpt_classes} -> "
+            f"Data classes from training data: {self.model_classes}",
+        )
+
+        for param_name in self.state_dict():
+            model_param = self.state_dict()[param_name].clone()
+            ckpt_param = state_dict[prefix + param_name]
+            if model_param.shape != ckpt_param.shape:
+                for model_t, ckpt_t in enumerate(model2ckpt):
+                    if ckpt_t >= 0:
+                        model_param[model_t].copy_(ckpt_param[ckpt_t])
+
+                # Replace checkpoint weight by mixed weights
+                state_dict[prefix + param_name] = model_param
+
+    @staticmethod
+    def map_class_names(src_classes: list[str], dst_classes: list[str]) -> list[int]:
+        """Computes src to dst index mapping.
+
+        src2dst[src_idx] = dst_idx
+        #  according to class name matching, -1 for non-matched ones
+        assert(len(src2dst) == len(src_classes))
+        ex)
+          src_classes = ['person', 'car', 'tree']
+          dst_classes = ['tree', 'person', 'sky', 'ball']
+          -> Returns src2dst = [1, -1, 0]
+        """
+        src2dst = []
+        for src_class in src_classes:
+            if src_class in dst_classes:
+                src2dst.append(dst_classes.index(src_class))
+            else:
+                src2dst.append(-1)
+        return src2dst
